+++ src/int/ft_intlen.c
@@ -6,13 +6,13 @@
 /*   By: msloot <msloot@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/15 12:29:36 by msloot            #+#    #+#             */
-/*   Updated: 2023/11/18 17:51:58 by msloot           ###   ########.fr       */
+/*   Updated: 2023/11/18 18:45:47 by msloot           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_intlen(long n)
+size_t	ft_intlen(ssize_t n)
 {
 	return (ft_intlen_base(n, 10));
 }
